Loader to no longer expose its ObjectifyImpl instance
- make sure that neither ofy nor the transactor are exposed and accessible where they shouldn't be
- expose only what is really needed - factory() and getCache()
