+++ src/main/java/com/googlecode/objectify/impl/LoaderImpl.java
@@ -6,6 +6,7 @@ import com.google.common.collect.Maps;
 import com.googlecode.objectify.Key;
 import com.googlecode.objectify.LoadResult;
 import com.googlecode.objectify.Objectify;
+import com.googlecode.objectify.ObjectifyFactory;
 import com.googlecode.objectify.Ref;
 import com.googlecode.objectify.Result;
 import com.googlecode.objectify.cmd.LoadType;
@@ -33,8 +34,8 @@ import java.util.Set;
 public class LoaderImpl<L extends Loader> extends Queryable<Object> implements Loader, Cloneable
 {
 	/** */
-	protected ObjectifyImpl<?> ofy;
-	protected Transactor<?> transactor;
+	private ObjectifyImpl<?> ofy;
+	private Transactor<?> transactor;
 
 	/** */
 	protected LoadArrangement loadArrangement = new LoadArrangement();
@@ -123,7 +124,7 @@ public class LoaderImpl<L extends Loader> extends Queryable<Object> implements L
 	 */
 	@Override
 	public <E> LoadResult<E> entity(E entity) {
-		return key(ofy.factory().keys().keyOf(entity));
+		return key(factory().keys().keyOf(entity));
 	}
 
 	/* (non-Javadoc)
@@ -132,7 +133,7 @@ public class LoaderImpl<L extends Loader> extends Queryable<Object> implements L
 	@Override
 	@SuppressWarnings("unchecked")
 	public <E> LoadResult<E> value(Object key) {
-		return (LoadResult<E>)key(ofy.factory().keys().anythingToKey(key));
+		return (LoadResult<E>)key(factory().keys().anythingToKey(key));
 	}
 
 	/* (non-Javadoc)
@@ -186,7 +187,7 @@ public class LoaderImpl<L extends Loader> extends Queryable<Object> implements L
 		// Do this in a separate pass so any errors converting keys will show up before we try loading something
 		List<Key<E>> keys = new ArrayList<>();
 		for (Object keyish: values)
-			keys.add((Key<E>)ofy.factory().keys().anythingToKey(keyish));
+			keys.add((Key<E>)factory().keys().anythingToKey(keyish));
 
 		LoadEngine engine = createLoadEngine();
 
@@ -212,14 +213,6 @@ public class LoaderImpl<L extends Loader> extends Queryable<Object> implements L
 	}
 
 	/* (non-Javadoc)
-	 * @see com.googlecode.objectify.cmd.Loader#getObjectify()
-	 */
-	@Override
-	public Objectify getObjectify() {
-		return ofy;
-	}
-
-	/* (non-Javadoc)
 	 * @see com.googlecode.objectify.cmd.Loader#getLoadGroups()
 	 */
 	@Override
@@ -227,9 +220,14 @@ public class LoaderImpl<L extends Loader> extends Queryable<Object> implements L
 		return Collections.unmodifiableSet(loadArrangement);
 	}
 
-	/** */
-	public ObjectifyImpl<?> getObjectifyImpl() {
-		return this.ofy;
+	/* Expose the factory but not expose ofy directly. */
+	public ObjectifyFactory factory() {
+		return ofy.factory();
+	}
+
+	/* Expose the cache setting but not expose ofy directly. */
+	public boolean getCache() {
+		return ofy.getCache();
 	}
 
 	/**
